# Refs #27624 -- Optimized Query.clone() for non-combined queries.